refactor: remove deprecated codes
