--- core/framework/mcp/__init__.py
@@ -1,4 +1,0 @@
-"""MCP servers for worker-bee."""
-
-# Don't auto-import servers to avoid double-import issues when running with -m
-__all__ = []
