+++ core/tests/test_mcp_server.py
@@ -42,14 +42,3 @@ class TestMCPDependencies:
         assert FastMCP is not None
 
 
-class TestMCPPackageExports:
-    """Tests for the framework.mcp package exports."""
-
-    def test_package_importable(self):
-        """Test that framework.mcp package can be imported."""
-        if not MCP_AVAILABLE:
-            pytest.skip(MCP_SKIP_REASON)
-
-        import framework.mcp
-
-        assert framework.mcp is not None
